refactor : dayjs 함수들을 util로 만듬
