+++ src/util/days.ts
@@ -1,0 +1,9 @@
+import dayjs from "dayjs"
+
+export const isCurrentDateAfter = (date: string) => {
+  return dayjs().isAfter(dayjs(date))
+}
+
+export const formatToDate = (date: string, format: string) => {
+  return dayjs(date).format(format)
+}
